Split incidents & mine_incidents at API level
