+++ frontend/src/constants/API.js
@@ -99,6 +99,6 @@ export const MINE_INCIDENT = (mineGuid, mine_incident_guid) =>
   `/mines/${mineGuid}/incidents/${mine_incident_guid}`;
 export const MINE_INCIDENT_DOCUMENT = (mineGuid) => `/mines/${mineGuid}/incidents/documents`;
 
-export const MINE_INCIDENT_FOLLOWUP_ACTIONS = `/mines/incidents/followup-types`;
-export const MINE_INCIDENT_DETERMINATION_TYPES = `/mines/incidents/determination-types`;
-export const MINE_INCIDENT_STATUS_CODES = `/mines/incidents/status-codes`;
+export const INCIDENT_FOLLOWUP_ACTIONS = `/incidents/followup-types`;
+export const INCIDENT_DETERMINATION_TYPES = `/incidents/determination-types`;
+export const INCIDENT_STATUS_CODES = `/incidents/status-codes`;
