+++ python-backend/app/api/incidents/models/mine_incident.py
@@ -5,12 +5,14 @@ from sqlalchemy.orm import validates
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.schema import FetchedValue
 from app.extensions import db
-from ....utils.models_mixins import AuditMixin, Base
+
 from .mine_incident_followup_investigation_type import MineIncidentFollowupInvestigationType
-from app.api.mines.incidents.models.mine_incident_determination_type import MineIncidentDeterminationType
+
+from app.api.utils.models_mixins import AuditMixin, Base
+from app.api.incidents.models.mine_incident_determination_type import MineIncidentDeterminationType
+from app.api.incidents.models.mine_incident_do_subparagraph import MineIncidentDoSubparagraph
+from app.api.incidents.models.mine_incident_recommendation import MineIncidentRecommendation
 from app.api.mines.compliance.models.compliance_article import ComplianceArticle
-from app.api.mines.incidents.models.mine_incident_do_subparagraph import MineIncidentDoSubparagraph
-from app.api.mines.incidents.models.mine_incident_recommendation import MineIncidentRecommendation
 
 
 class MineIncident(AuditMixin, Base):
