+++ python-backend/app/api/incidents/models/mine_incident_determination_type.py
@@ -1,6 +1,7 @@
 from sqlalchemy.schema import FetchedValue
 from app.extensions import db
-from ....utils.models_mixins import AuditMixin, Base
+
+from app.api.utils.models_mixins import AuditMixin, Base
 
 
 class MineIncidentDeterminationType(AuditMixin, Base):
