+++ python-backend/app/api/incidents/models/mine_incident_do_subparagraph.py
@@ -1,5 +1,6 @@
 from app.extensions import db
-from ....utils.models_mixins import Base
+
+from app.api.utils.models_mixins import Base
 
 
 class MineIncidentDoSubparagraph(Base):
