+++ python-backend/app/api/incidents/models/mine_incident_recommendation.py
@@ -5,7 +5,8 @@ from sqlalchemy.orm import validates
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.schema import FetchedValue
 from app.extensions import db
-from ....utils.models_mixins import AuditMixin, Base
+
+from app.api.utils.models_mixins import AuditMixin, Base
 
 
 class MineIncidentRecommendation(AuditMixin, Base):
