+++ python-backend/app/api/incidents/resources/mine_incident_followup_types.py
@@ -7,8 +7,8 @@ from app.extensions import api
 from app.api.utils.resources_mixins import UserMixin
 from app.api.utils.access_decorators import requires_role_view_all, requires_role_mine_edit
 
-from ..models.mine_incident_followup_investigation_type import MineIncidentFollowupInvestigationType
-from ...mine_api_models import MINE_INCIDENT_FOLLOWUP_INVESTIGATION_TYPE_MODEL
+from app.api.incidents.response_models import MINE_INCIDENT_FOLLOWUP_INVESTIGATION_TYPE_MODEL
+from app.api.incidents.models.mine_incident_followup_investigation_type import MineIncidentFollowupInvestigationType
 
 
 class MineIncidentFollowupTypeResource(Resource, UserMixin):
