+++ python-backend/app/api/mines/namespace/mines.py
@@ -20,9 +20,6 @@ from ..variances.resources.variance_list import MineVarianceListResource
 from ..variances.resources.variance_document_upload import MineVarianceDocumentUploadResource
 from ..variances.resources.variance_uploaded_documents import MineVarianceUploadedDocumentsResource
 from ..incidents.resources.mine_incidents import MineIncidentListResource, MineIncidentResource
-from ..incidents.resources.mine_incident_status_codes import MineIncidentStatusCodeResource
-from ..incidents.resources.mine_incident_followup_types import MineIncidentFollowupTypeResource
-from ..incidents.resources.mine_incident_determination_types import MineIncidentDeterminationTypeResource
 from ..incidents.resources.mine_incident_document import MineIncidentDocumentListResource, MineIncidentDocumentResource
 from app.api.mines.permits.permit.resources.permit import PermitResource, PermitListResource
 from app.api.mines.permits.permit.resources.permit_status_code import PermitStatusCodeResource
@@ -73,10 +70,6 @@ api.add_resource(
 
 api.add_resource(MineIncidentListResource, '/<string:mine_guid>/incidents')
 api.add_resource(MineIncidentResource, '/<string:mine_guid>/incidents/<string:mine_incident_guid>')
-
-api.add_resource(MineIncidentFollowupTypeResource, '/incidents/followup-types')
-api.add_resource(MineIncidentDeterminationTypeResource, '/incidents/determination-types')
-api.add_resource(MineIncidentStatusCodeResource, '/incidents/status-codes')
 
 api.add_resource(
     MineIncidentDocumentResource,
